Avoid non-portable %zu.
git-svn-id: http://svn.drobilla.net/sord/trunk@196 3d64ff67-21c5-427c-a301-fe4f08042e5a

+++ src/sordi.c
@@ -167,7 +167,8 @@ main(int argc, char** argv)
 
 	serd_reader_free(reader);
 
-	fprintf(stderr, "Loaded %zu statements\n", sord_num_quads(sord));
+	fprintf(stderr, "Loaded %lu statements\n",
+	        (unsigned long)sord_num_quads(sord));
 
 	SerdEnv* write_env = serd_env_new(&base_uri_node);
